fix: fixed back button inside the articles
