+++ components/layout/Footer.tsx
@@ -33,8 +33,8 @@ export function Footer() {
                 <NavLink href='/uses'>Uses</NavLink>
               </div>
               <p className='text-sm text-zinc-400 dark:text-zinc-500'>
-                &copy; {new Date().getFullYear()} Spencer Sharp. All rights
-                reserved.
+                Made with ☕️ and ❤️ by Azaz Ahamed. &copy;{' '}
+                {new Date().getFullYear()} All rights reserved.
               </p>
             </div>
           </Container.Inner>
